Move funsor.source -> funsor.syntax
