+++ funsor/syntax.py
